Createcomment

+++ app/controllers/artist_favorites_controller.rb
@@ -1,17 +1,16 @@
 class ArtistFavoritesController < ApplicationController
   before_action :authenticate_user!
   def create
-    artist=Artist.find(params[:every_artist_id])
-    favorite=ArtistFavorite.new(artist_id: artist.id)
-    favorite.user_id=current_user.id
-    favorite.save
-    redirect_to every_artist_path(artist.id)
-  end
-
-  def destroy
-    artist=Artist.find(params[:every_artist_id])
-    favorite=current_user.artist_favorites.find_by(artist_id: artist.id)
-    favorite.destroy
-    redirect_to every_artist_path(artist.id)
+    @artist=Artist.find(params[:every_artist_id])
+     if @artist.favorited_by?(current_user)
+      favorite=current_user.artist_favorites.find_by(artist_id: @artist.id)
+      favorite.destroy
+      render json: @artist.id
+    else
+      favorite=ArtistFavorite.new(artist_id: @artist.id)
+      favorite.user_id=current_user.id
+      favorite.save
+      render json: @artist.id
+    end
   end
 end
